Remove previousPath related logic.
Remove the previousPath reducer and related conditional `router.goBack` logic. This never really worked correctly in practice since using the back function of history leaves the reducer with the wrong `previousPath` value (actually the "next" path). A proper solution would be to store an array of location history, and push or pop the location on navigation.

It's better to just keep things simple.