Refactor connector config and introduce scope.
All of the `relocation` connector config can be assigned as props on the connected component instance or configured within the call to the `relocation` connector.

Instead of taking separate render map and config parameters, the connector now takes a single default props object. The render map can be defined within as `components` or can be provided as a `components` prop on the rendered instance.

In addition an optional `scope` option allows namespacing the consumed and generated props. This is designed to prevent conflicts with existing props on the connected component and allows composition with multiple `relocation` connector calls on the same component.