Remove `context`, `defaultProps`, add `componentShouldMount`.
The previous `context` property was basically used to isolate slices of `redux` for a particular component tree and `defaultProps` was used to pass in `components` as a convenience method. The `context` property can be emulated by either using `getRelocationState` or `componentShouldMount` now, and `defaultProps` just makes things messy internally, and one should just use `withProps` from `recompose`.

The new `componentShouldMount` is a mechanism for determining if a given component from the `redux` store should be passed through to the component decorated with `relocation()`. It has the signature `(state, props, component)` which mimics the `mapStateToProps` function of `redux` except that the third argument is the component from the store. Note that the `props` here come from the decorated component. This allows you much better control of determining what components from the large list of components in the store actually make it down.